Don't use gtid-strict-mode in mariadb-binlog restore
Don't use gtid-strict-mode in mariadb-binlog. It's enabled by default in
mariadb-binlog, but disabled in mariadb-server, so having this option
enabled leads to the following error, when replaying from multiple (>1)
binlog files:

```
ERROR: Binary logs are missing data for domain 0.
Expected data to start from state 0-0-0;
however, the initial GTID state of the logs was 0-1-406690.
```